- Fix incorrect function definitions. To be submitted upstream.
svn path=/branches/arwinss/; revision=48380

+++ reactos/dll/win32/gdi32/bitmap.c
@@ -433,7 +433,7 @@ LONG WINAPI GetBitmapBits(
  */
 LONG WINAPI SetBitmapBits(
     HBITMAP hbitmap, /* [in] Handle to bitmap */
-    LONG count,        /* [in] Number of bytes in bitmap array */
+    DWORD count,       /* [in] Number of bytes in bitmap array */
     LPCVOID bits)      /* [in] Address of array with bitmap bits */
 {
     BITMAPOBJ *bmp;
